#63 - refactor: 회원 Service 단위 테스트 단일 책임 원칙 적용
